Small tweaks
Remove unnecessaries parentheses, casts, semi-colons, etc

+++ src/Faker/Provider/lt_LT/Person.php
@@ -302,7 +302,7 @@ class Person extends \Faker\Provider\Person
             $birthdate = \Faker\Provider\DateTime::dateTimeThisCentury();
         }
 
-        $genderNumber = ($gender == 'male') ? (int) 1 : (int) 0;
+        $genderNumber = ($gender == 'male') ? 1 : 0;
         $firstNumber = (int) floor($birthdate->format('Y') / 100) * 2 - 34 - $genderNumber;
 
         $datePart = $birthdate->format('ymd');
@@ -318,7 +318,7 @@ class Person extends \Faker\Provider\Person
         }
 
         $sum = self::calculateSum($partOfPerosnalCode, 2);
-        $liekana = (int) $sum % 11;
+        $liekana = $sum % 11;
 
         $lastNumber = ($liekana !== 10) ? $liekana : 0;
         return $firstNumber . $datePart . $randomDigits . $lastNumber;
